requirements.txt
django==3.1.6
djangorestframework==3.12.2
markdown==3.3.3
jwt==1.11.0